+++ app/src/main/java/org/rdr/radarbox/Plots2D/Line2D.java
@@ -4,13 +4,13 @@ import java.io.Serializable;
 import java.util.Arrays;
 
 public class Line2D implements Serializable {
-    private double[] x;
-    private double[] y;
+    private float[] x;
+    private float[] y;
     private int color; //заменить на PAINT
     private String legendName;
     private boolean needShow;
 
-    public Line2D(double[] x, double[] y, int color, String legendName) {
+    public Line2D(float[] x, float[] y, int color, String legendName) {
         this.x = x;
         this.y = y;
         this.color = color;
@@ -18,14 +18,14 @@ public class Line2D implements Serializable {
         this.needShow = true;
     }
 
-    public void setX(double[] x)           { this.x= Arrays.copyOf(x,x.length);}
-    public void setY(double[] y)           {this.y= Arrays.copyOf(y,y.length);}
+    public void setX(float[] x)           { this.x= Arrays.copyOf(x,x.length);}
+    public void setY(float[] y)           {this.y= Arrays.copyOf(y,y.length);}
     void setColor(int c)          {this.color=c;}
     void setLegendName(String s)    {this.legendName=s;}
     void setNeedShow(boolean value) {this.needShow = value;}
 
-    public double[] getX()         {return this.x;}
-    double[] getY()         {return this.y;}
+    public float[] getX()         {return this.x;}
+    float[] getY()         {return this.y;}
     int getColor()        {return this.color;}
     String getLegendName()  {return this.legendName;}
     boolean isNeedShow()    {return this.needShow;}
